remove demo and make test internal
