--- streams/examples/src/main/java/org/apache/kafka/streams/examples/wordcount/WindowedWordCountProcessorDemo.java
@@ -1,155 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.kafka.streams.examples.wordcount;
-
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.common.serialization.Serdes;
-import org.apache.kafka.streams.KafkaStreams;
-import org.apache.kafka.streams.KeyValue;
-import org.apache.kafka.streams.StreamsConfig;
-import org.apache.kafka.streams.Topology;
-import org.apache.kafka.streams.kstream.Windowed;
-import org.apache.kafka.streams.processor.Processor;
-import org.apache.kafka.streams.processor.ProcessorContext;
-import org.apache.kafka.streams.processor.ProcessorSupplier;
-import org.apache.kafka.streams.processor.PunctuationType;
-import org.apache.kafka.streams.state.KeyValueIterator;
-import org.apache.kafka.streams.state.Stores;
-import org.apache.kafka.streams.state.WindowStore;
-
-import java.time.Duration;
-import java.util.Locale;
-import java.util.Properties;
-import java.util.concurrent.CountDownLatch;
-
-/**
- * Demonstrates, using the low-level Processor APIs, how to implement the WordCount program
- * that computes a simple word occurrence histogram from an input text. Also demonstrates the
- * use of specific store implementations, in this case, a windowed store.
- * <p>
- * <strong>Note: This is simplified code that only works correctly for single partition input topics.</strong>
- * <p>
- * Before running this example you must create the input topic and the output topic (e.g. via
- * {@code bin/kafka-topics.sh --create ...}), and write some data to the input topic (e.g. via
- * {@code bin/kafka-console-producer.sh}). Otherwise you won't see any data arriving in the output topic.
- */
-public final class WindowedWordCountProcessorDemo {
-
-    private WindowedWordCountProcessorDemo() {}
-
-    static class MyProcessorSupplier implements ProcessorSupplier<String, String> {
-
-        @Override
-        public Processor<String, String> get() {
-            return new Processor<String, String>() {
-                private ProcessorContext context;
-                private WindowStore<String, Integer> windowStore;
-
-                @Override
-                @SuppressWarnings("unchecked")
-                public void init(final ProcessorContext context) {
-                    this.context = context;
-                    this.context.schedule(Duration.ofSeconds(1), PunctuationType.STREAM_TIME, timestamp -> {
-                        try (final KeyValueIterator<Windowed<String>, Integer> iter = windowStore.all()) {
-                            System.out.println("----------- " + timestamp + " ----------- ");
-
-                            while (iter.hasNext()) {
-                                final KeyValue<Windowed<String>, Integer> entry = iter.next();
-
-                                System.out.println("[" + entry.key + ", " + entry.value + "]");
-
-                                context.forward(entry.key.toString(), entry.value.toString());
-                            }
-                        }
-                    });
-                    this.windowStore = (WindowStore<String, Integer>) context.getStateStore("WindowedCounts");
-                }
-
-                @Override
-                public void process(final String dummy, final String line) {
-                    final String[] words = line.toLowerCase(Locale.getDefault()).split(" ");
-                    final long timestamp = this.context.timestamp();
-
-                    // calculate the window as every 100 ms
-                    // Note this has to be aligned with the configuratoin for the window store you register separately
-                    final long windowStart = timestamp / 100 * 100;
-
-                    for (final String word : words) {
-                        final Integer oldValue = this.windowStore.fetch(word, windowStart);
-
-                        if (oldValue == null) {
-                            this.windowStore.put(word, 1, windowStart);
-                        } else {
-                            this.windowStore.put(word, oldValue + 1, windowStart);
-                        }
-                    }
-                }
-
-                @Override
-                public void close() {}
-            };
-        }
-    }
-
-    public static void main(final String[] args) {
-        final Properties props = new Properties();
-        props.put(StreamsConfig.APPLICATION_ID_CONFIG, "streams-wordcount-processor");
-        props.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
-        props.put(StreamsConfig.CACHE_MAX_BYTES_BUFFERING_CONFIG, 0);
-        props.put(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, Serdes.String().getClass());
-        props.put(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, Serdes.String().getClass());
-
-        // setting offset reset to earliest so that we can re-run the demo code with the same pre-loaded data
-        props.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
-
-        final Topology builder = new Topology();
-
-        builder.addSource("Source", "streams-plaintext-input");
-
-        builder.addProcessor("Process", new MyProcessorSupplier(), "Source");
-        builder.addStateStore(
-            Stores.windowStoreBuilder(
-                Stores.inMemoryWindowStore("WindowedCounts", Duration.ofHours(1), Duration.ofMillis(100), false),
-                Serdes.String(),
-                Serdes.Integer()
-            ),
-            "Process"
-        );
-
-        builder.addSink("Sink", "streams-wordcount-processor-output", "Process");
-
-        final KafkaStreams streams = new KafkaStreams(builder, props);
-        final CountDownLatch latch = new CountDownLatch(1);
-
-        // attach shutdown handler to catch control-c
-        Runtime.getRuntime().addShutdownHook(new Thread("streams-wordcount-shutdown-hook") {
-            @Override
-            public void run() {
-                streams.close();
-                latch.countDown();
-            }
-        });
-
-        try {
-            streams.start();
-            latch.await();
-        } catch (final Throwable e) {
-            System.exit(1);
-        }
-        System.exit(0);
-    }
-}
